+++ frontend/src/pages/MainNotice/components/ClipModal/ClipModal.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { AiOutlineClose } from 'react-icons/ai';
 import styles from './ClipModal.module.scss';
-import { addScraps } from "../../api/userAPI";
+import { addScraps } from "../../../../api/noticeAPI";
 
 const ClipModal = ({ noticeId, categoryId, onModalClose, openEventModal, onScrapsData }) => {
