+++ frontend/src/pages/MainNotice/components/ClipModal/ClipModal.module.scss
