+++ frontend/src/pages/MainNotice/components/NoticeItem/NoticeItem.module.scss
