+++ frontend/src/pages/MainNotice/components/UserCategory/UserCategory.module.scss
